Projets/Projet P4 - Triof/une-solution/requirements.txt
flask
matplotlib
numpy
pillow
requests
scikit-image
tflite-runtime